reorganize shortcut files
