+++ src/listeners/player/shortcuts/playback-rate.js
@@ -1,5 +1,5 @@
-var player = require('../config/youtube.js')
-var editor = require('../config/editor.js')
+var player = require('../../../config/youtube.js')
+var editor = require('../../../config/editor.js')
 
 function adjustRate (event) {
   if ( editor.status && !editor.areaSelect ) {
