+++ src/listeners/shortcuts/note-creator.js
@@ -1,6 +1,6 @@
-import { createNote } from '../stage/elements/create-note.js'
-var editor = require('../config/editor.js')
-var player = require('../config/youtube.js')
+import { createNote } from '../../stage/elements/create-note.js'
+var editor = require('../../config/editor.js')
+var player = require('../../config/youtube.js')
 
 function createNewNote (event) {
   if ( editor.status && player.getState() === 'paused' && !editor.areaSelect ) {
